fix: secure E2B sandbox traffic
